Correcting the typescript test
In the last test, the call to withdraw wasn't executed by the "secondAccount" so the balance stayed the same before and after the call to "withdraw". 
And withdraw doesn't take any argument (as in the Javascript version).
Also, gas wasn't taken into account in the expected balance at the end.
I corrected those 3 problems.